Terminación de Carga y Descarga de archivos

+++ database/migrations/2021_06_01_022369_create_reviews_table.php
@@ -17,7 +17,7 @@ class CreateReviewsTable extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained();
             $table->foreignId('pelicula_id')->constrained();
-            $table->float('valoracion', 2, 2);
+            $table->integer('valoracion');
             $table->text('resena'); 
             $table->timestamps();
         });
